+++ src/main/kotlin/org/jetbrains/intellij/platform/gradle/tasks/RunIdeTask.kt
@@ -13,6 +13,7 @@ import org.jetbrains.intellij.platform.gradle.Constants.Plugin
 import org.jetbrains.intellij.platform.gradle.Constants.Tasks
 import org.jetbrains.intellij.platform.gradle.tasks.aware.IntelliJPlatformVersionAware
 import org.jetbrains.intellij.platform.gradle.tasks.aware.RunnableIdeAware
+import org.jetbrains.intellij.platform.gradle.tasks.aware.SplitModeAware
 import org.jetbrains.intellij.platform.gradle.utils.asPath
 import kotlin.io.path.pathString
 
@@ -24,7 +25,7 @@ import kotlin.io.path.pathString
  * To register a customized task, use [CustomRunIdeTask] instead.
  */
 @UntrackedTask(because = "Should always run")
-abstract class RunIdeTask : JavaExec(), RunnableIdeAware, IntelliJPlatformVersionAware {
+abstract class RunIdeTask : JavaExec(), RunnableIdeAware, SplitModeAware, IntelliJPlatformVersionAware {
 
     init {
         group = Plugin.GROUP_NAME
@@ -43,9 +44,9 @@ abstract class RunIdeTask : JavaExec(), RunnableIdeAware, IntelliJPlatformVersio
 
         if (splitMode.get()) {
             environment("JETBRAINS_CLIENT_JDK", runtimeDirectory.asPath.pathString)
-            environment("JETBRAINS_CLIENT_PROPERTIES", frontendPropertiesFile.asPath.pathString)
+            environment("JETBRAINS_CLIENT_PROPERTIES", splitModeFrontendProperties.asPath.pathString)
 
-            if (args.orEmpty().isNotEmpty()) {
+            if (args.isNotEmpty()) {
                 throw InvalidUserDataException("Passing arguments directly is not supported in Split Mode. Use `argumentProviders` instead.")
             }
         }